refactor(core): split scale-style into collect, discrete, numeric, finite
Break the 871-line style scale resolver into family modules mirroring the
color split: thin resolveStyleScale orchestrator plus collect, shared
discrete training, numeric families, and finite (shape/linetype) families.
Add characterization tests for empty mapping, scaled constants, and
source-catalog ordinal stability under filtered frames.
