+++ packages/core/src/pipeline/scale-style-types.ts
@@ -1,9 +1,13 @@
-import type { StyleAesthetic } from "@ggsvelte/spec";
+import type { PortableSpec, StyleAesthetic } from "@ggsvelte/spec";
 
 import type { GuidePlan } from "../layout/temporal-guide.js";
 import type { LegendInput } from "../legend.js";
 import type { ResolvedStyleScale } from "../scales/style.js";
 import type { ScaleState } from "../scales/state.js";
+
+export type NumericStyleAesthetic = Extract<StyleAesthetic, "size" | "linewidth" | "alpha">;
+export type FiniteStyleAesthetic = Extract<StyleAesthetic, "shape" | "linetype">;
+export type FiniteStyleConfig = NonNullable<PortableSpec["scales"]>[FiniteStyleAesthetic];
 
 export interface StyleResolution {
   aesthetic: StyleAesthetic;
